created user table and completed user api

+++ api/classes.api.php
@@ -8,27 +8,35 @@ class Classes extends DatabaseConnection
     public function create_class_api($conn)
     {
         $response = [];
+        // check all $_POST if not empty
         if (!empty($_POST['name']) && !empty($_POST['description']) && !empty($_POST['s_id'])) {
             $name = $_POST['name'];
             $description = $_POST['description'];
             $s_id = $_POST['s_id'];
-            $query = "INSERT INTO `classes`(`name`, `description`,`s_id`) 
-        VALUES ('$name','$description','$s_id')";
-            $result = $conn->query($query);
-            if ($result) {
-                $response = ['status' => true, 'data' => 'successfully created'];
+
+            // Check if class exist
+            $check_query = "SELECT * FROM `classes` WHERE `name`='$name'";
+            $check_result = $conn->query($check_query);
+            if ($check_result && $check_result->num_rows > 0) {
+                $response = ['status' => false, 'data' => 'class name already exist'];
             } else {
-                $response = ['status' => true, 'data' => $conn->error];
+                $query = "INSERT INTO `classes`(`name`, `description`,`s_id`) 
+                VALUES ('$name','$description','$s_id')";
+                $result = $conn->query($query);
+                if ($result) {
+                    $response = ['status' => true, 'data' => 'successfully created'];
+                } else {
+                    $response = ['status' => false, 'data' => $conn->error];
+                }
             }
         } else {
-            $response = ['status' => true, 'data' => 'missing required fields'];
+            $response = ['status' => false, 'data' => 'missing required fields'];
         }
 
         echo json_encode($response);
     }
 
     // read all classes API
-
     public function read_all_class_api($conn)
     {
         $response = [];
@@ -38,8 +46,8 @@ class Classes extends DatabaseConnection
         $result = $conn->query($query);
 
         if ($result) {
-            while($row = $result->fetch_assoc()) {
-                $data []= $row;
+            while ($row = $result->fetch_assoc()) {
+                $data[] = $row;
             }
             $response = ['status' => true, 'data' => $data];
         } else {
@@ -58,7 +66,7 @@ class Classes extends DatabaseConnection
 
         if ($result) {
             while ($row = $result->fetch_assoc()) {
-                $data []= $row;
+                $data[] = $row;
             }
             $response = ['status' => true, 'data' => $data];
         } else {
@@ -74,14 +82,15 @@ class Classes extends DatabaseConnection
     {
         $response = [];
         $data = [];
+        // check  $_POST class ID  if not empty
         if (!empty($_POST['c_id'])) {
             $c_id = $_POST['c_id'];
             $query = "SELECT * FROM classes where c_id = '$c_id'";
             $result = $conn->query($query);
 
             if ($result) {
-                while($row = $result->fetch_assoc()) {
-                    $data []= $row;
+                while ($row = $result->fetch_assoc()) {
+                    $data[] = $row;
                 }
                 $response = ['status' => true, 'data' => $data];
             } else {
@@ -93,19 +102,26 @@ class Classes extends DatabaseConnection
         echo json_encode($response);
     }
 
-
+    // delete class API
     public function delete_class_api($conn)
     {
         $response = [];
+        // check  $_POST class ID  if not empty
         if (!empty($_POST['c_id'])) {
             $c_id = $_POST['c_id'];
-            $query = "DELETE FROM `classes` WHERE  c_id = '$c_id'";
-            $result = $conn->query($query);
-
-            if ($result) {
-                $response = ['status' => true, 'data' => 'successfully deleted'];
+            // Check if semester exists
+            $check_query = "SELECT * FROM `classes` WHERE `c_id`='$c_id'";
+            $check_result = $conn->query($check_query);
+            if ($check_result && $check_result->num_rows > 0) {
+                $query = "DELETE FROM `classes` WHERE  c_id = '$c_id'";
+                $result = $conn->query($query);
+                if ($result) {
+                    $response = ['status' => true, 'data' => 'successfully deleted'];
+                } else {
+                    $response = ['status' => false, 'data' => $conn->error];
+                }
             } else {
-                $response = ['status' => false, 'data' => $conn->error];
+                $response = ['status' => false, 'data' => 'class does not exist'];
             }
         } else {
             $response = ['status' => false, 'data' => 'missing required field'];
@@ -117,8 +133,11 @@ class Classes extends DatabaseConnection
     public function update_class_api($conn)
     {
         $response = [];
-        if (!empty($_POST['s_id']) && !empty($_POST['name']) && !empty($_POST['description']) &&
-        !empty($_POST['c_id'])  ) {
+        // check all $_POST if not empty
+        if (
+            !empty($_POST['s_id']) && !empty($_POST['name']) && !empty($_POST['description']) &&
+            !empty($_POST['c_id'])
+        ) {
             $s_id = $_POST['s_id'];
             $c_id = $_POST['c_id'];
             $name = $_POST['name'];
@@ -156,6 +175,6 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     } else {
         echo json_encode(['status' => false, 'data' => 'Action is required']);
     }
-}else {
+} else {
     echo json_encode(['status' => false, 'data' => 'Invalid request method']);
 }
